+++ app/api/deleteUserCartByPidUserEmail/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function DELETE(req: NextRequest) {
   try {
@@ -16,7 +16,7 @@ export async function DELETE(req: NextRequest) {
 
     product_id = parseInt(product_id);
 
-    await helpers.deleteUserCartByPidUserEmail(user_email, product_id);
+    await db.deleteUserCartByPidUserEmail(user_email, product_id);
 
     return NextResponse.json(
       { msg: "Item removed from user cart successfully!" },
