+++ app/api/deleteVendorRequestByUserEmail/route.ts
@@ -1,12 +1,12 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function DELETE(req: NextRequest) {
   try {
     const { user_email } = await req.json();
 
     // Call the helper function to delete the vendor request by user email
-    await helpers.deleteVendorRequestByUserEmail(user_email);
+    await db.deleteVendorRequestByUserEmail(user_email);
     console.log("Vendor Request Deleted Successfully!");
 
     return NextResponse.json({}, { status: 200 });
