+++ app/api/getAllWarehouseInfo/route.ts
@@ -1,10 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function GET(req: NextRequest) {
   try {
     // Fetch all warehouse info using the helper function
-    const response = await helpers.getAllWarehouseInfo();
+    const response = await db.getAllWarehouseInfo();
 
     return NextResponse.json(response, { status: 200 });
   } catch (error) {
