+++ app/api/getUserCartByUserEmail/[user_email]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../../db");
+const db = require("@/app/api/db");
 
 export async function GET(
   req: NextRequest,
@@ -12,7 +12,7 @@ export async function GET(
   const user_email = params.user_email.trim();
 
   try {
-    const products = await helpers.getUserCartByUserEmail(user_email);
+    const products = await db.getUserCartByUserEmail(user_email);
     return NextResponse.json(products, { status: 200 });
   } catch (error) {
     console.error("Error:", error);
