+++ app/api/getUserWishlistByUserEmail/[user_email]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../../db");
+const db = require("@/app/api/db");
 
 export async function GET(
   req: NextRequest,
@@ -12,7 +12,7 @@ export async function GET(
   const user_email = params.user_email.trim();
 
   try {
-    const products = await helpers.getUserWishlistByUserEmail(user_email);
+    const products = await db.getUserWishlistByUserEmail(user_email);
     return NextResponse.json(products, { status: 200 });
   } catch (error) {
     console.error("Error:", error);
