+++ app/api/getWarehouseInfo/[warehouse_id]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../../db");
+const db = require("@/app/api/db");
 
 export async function GET(
   req: NextRequest,
@@ -9,7 +9,7 @@ export async function GET(
     const { warehouse_id } = params;
 
     // Fetch the warehouse info based on the warehouse_id
-    const response = await helpers.getWarehouseInfo(warehouse_id);
+    const response = await db.getWarehouseInfo(warehouse_id);
 
     return NextResponse.json(response, { status: 200 });
   } catch (error) {
