+++ app/api/patchUserAddress/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function PATCH(req: NextRequest) {
   try {
@@ -45,7 +45,7 @@ export async function PATCH(req: NextRequest) {
     }
     country = country.trim();
 
-    await helpers.patchUserAddress(
+    await db.patchUserAddress(
       user_email,
       street_name,
       city,
