+++ app/api/patchUserType/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function PATCH(req: NextRequest) {
   try {
@@ -25,7 +25,7 @@ export async function PATCH(req: NextRequest) {
 
     const type_id = user_type === "vendor" ? 1 : 2;
 
-    await helpers.patchUserType(user_email, type_id);
+    await db.patchUserType(user_email, type_id);
 
     return NextResponse.json(
       { success: "User type modified successfully!" },
