+++ app/api/updateProductPriceByProductId/route.ts
@@ -1,12 +1,12 @@
 import { NextRequest, NextResponse } from "next/server";
-const { helpers } = require("../db");
+const db = require("@/app/api/db");
 
 export async function PATCH(req: NextRequest) {
   try {
     const { product_id, new_price } = await req.json();
 
     // Call the helper function to update the product price
-    await helpers.updateProductPriceByProductId(product_id, new_price);
+    await db.updateProductPriceByProductId(product_id, new_price);
 
     return NextResponse.json({}, { status: 200 });
   } catch (error) {
